Moved resetTracklist to sendUrlPostRequest to clean up file structure. #4
